% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_diffusion_matrices.R
\docType{data}
\name{diffusion_matrix}
\alias{diffusion_matrix}
\title{Named list of standard error diffusion matrices}
\format{An object of class \code{list} of length 10.}
\usage{
diffusion_matrix
}
\description{
Named list of standard error diffusion matrices
}
\keyword{datasets}
