% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dither.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Add a border of zeros around a matrix}
\usage{
expand_matrix(m, border)
}
\arguments{
\item{m}{matrix}

\item{border}{border width}
}
\value{
a copy of the matrix embedded within a border of zeros
}
\description{
Add a border of zeros around a matrix
}
