% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dither.R
\name{view_matrix}
\alias{view_matrix}
\title{Plot a matrix as an image}
\usage{
view_matrix(m)
}
\arguments{
\item{m}{numeric matrix with all values in the range [0, 1]}
}
\description{
Plot a matrix as an image
}
